[Intermediate] Base version of OLC parser -copy paste extracted from OLC
